pyproject.toml
[tool.poetry]
name = "primal"
version = "0.1.0"
description = "PRIMAL"
authors = ["Yan <yan@meshcapade.com>"]
readme = "README.md"
packages = [{include = "primal"}]



[tool.poetry.dependencies]
python = ">=3.10"
hydra-core = ">=1.3.2"
lightning = ">=2.1.3"
human-body-prior = {git = "https://github.com/nghorbani/human_body_prior.git"}
numpy = ">=1.21.0"
tensorboard = ">=2.15.1"
hydra-colorlog = ">=1.2.0"
torchvision = ">=0.16.2"
opencv-python = ">=4.9.0.80"
diffusers = ">=0.26.3"
accelerate = ">=0.27.2"
transformers = ">=4.38.2"
maturin = ">=1.7.0"
einops = ">=0.8.0"
gradio = ">=4.40.0"
gdown = ">=5.2.0"
smplx = ">=0.1.28"
tyro = ">=0.7.0"

huggingface-hub = ">=0.25.0"
keyboard = ">=0.13.5"
sshkeyboard = ">=2.3.1"
pendulum = ">=3.0.0"
smplcodec = ">=2.0.2"

trimesh = "^4.6.10"
pynput = "^1.8.1"
multidict = "^6.6.4"
gloss-rs = "^0.7.0"
smpl-rs = "^0.7.0"
python-dotenv = "^1.1.1"



[tool.poetry.group.dev.dependencies]
ipdb = ">=0.13.13"
matplotlib = ">=3.8.2"



[tool.poetry.group.contacts.dependencies]
juliacall = ">=0.9.15"
scikit-learn = ">=1.4.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
